relay: updating the embedded sdk
